+++ src/services/twitter/article.rs
@@ -156,8 +156,8 @@ impl ArticleData for TweetArticleData {
 	fn id(&self) -> String {
 		self.id.to_string()
 	}
-	fn sortable_id(&self) -> usize {
-		self.id as usize
+	fn sortable_id(&self) -> u64 {
+		self.id as u64
 	}
 	fn creation_time(&self) -> Date {
 		self.creation_time.clone()
